requirements.txt
python-telegram-bot==20.7
Pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-dotenv==1.0.0
loguru==0.7.2
